+++ tensordict/base.py
@@ -61,6 +61,7 @@ from tensordict.utils import (
     _lock_warn,
     _make_dtype_promotion,
     _parse_to,
+    _pin_mem,
     _prefix_last_key,
     _proc_init,
     _prune_selected_keys,
@@ -3625,6 +3626,7 @@ class TensorDictBase(MutableMapping):
         )
 
     def is_consolidated(self):
+        """Checks if a TensorDict has a consolidated storage."""
         return hasattr(self, "_consolidated")
 
     def memmap_(
@@ -9728,15 +9730,3 @@ def _expand_to_match_shape(
         batch_size = torch.Size([*parent_batch_size, *_shape(data)[self_batch_dims:]])
         result = data.empty(batch_size=batch_size)
     return result
-
-
-def _pin_mem(q_in, q_out):
-    while not q_in.empty():
-        input = q_in.get(timeout=_PIN_MEM_TIMEOUT)
-        try:
-            key, val = input[0], input[1].pin_memory()
-        except Exception as err:
-            # Surface the exception
-            q_out.put(err)
-            return
-        q_out.put((key, val))
